Refactored: created a VoiceChat base namespace and a Legacy Network namespace
