+++ VoiceChat/Assets/VoiceChat/Scene/LegacyNetworking.unity
@@ -150,6 +150,7 @@ Camera:
   m_OcclusionCulling: 1
   m_StereoConvergence: 10
   m_StereoSeparation: .0219999999
+  m_StereoMirrorMode: 0
 --- !u!81 &718837315
 AudioListener:
   m_ObjectHideFlags: 0
