+++ VoiceChat/Assets/VoiceChat/Scripts/Demo/VoiceChatPickUi.cs
@@ -1,29 +1,32 @@
 using UnityEngine;
 using System.Collections;
 
-public class VoiceChatPickUi : MonoBehaviour
+namespace VoiceChat.Demo.LegacyNetworking
 {
-    void OnGUI()
+    public class VoiceChatPickUi : MonoBehaviour
     {
-        GUILayout.Window(0, new Rect(Screen.width / 2 - 50, Screen.height / 2 - 50, 100, 100), Window, "", GUIStyle.none);
-    }
-
-    void Window(int id)
-    {
-        GUI.Box(new Rect(0, 0, 100, 100), "");
-
-        if (GUILayout.Button("Start Server"))
+        void OnGUI()
         {
-            gameObject.GetComponent<VoiceChatUnityServer>().enabled = true;
-            gameObject.AddComponent<VoiceChatServerUi>();
-            MonoBehaviour.Destroy(this);
+            GUILayout.Window(0, new Rect(Screen.width / 2 - 50, Screen.height / 2 - 50, 100, 100), Window, "", GUIStyle.none);
         }
 
-        if (GUILayout.Button("Start Client"))
+        void Window(int id)
         {
-            gameObject.GetComponent<VoiceChatUnityClient>().enabled = true;
-            gameObject.AddComponent<VoiceChatUi>();
-            MonoBehaviour.Destroy(this);
+            GUI.Box(new Rect(0, 0, 100, 100), "");
+
+            if (GUILayout.Button("Start Server"))
+            {
+                gameObject.GetComponent<VoiceChatUnityServer>().enabled = true;
+                gameObject.AddComponent<VoiceChatServerUi>();
+                MonoBehaviour.Destroy(this);
+            }
+
+            if (GUILayout.Button("Start Client"))
+            {
+                gameObject.GetComponent<VoiceChatUnityClient>().enabled = true;
+                gameObject.AddComponent<VoiceChatUi>();
+                MonoBehaviour.Destroy(this);
+            }
         }
-    }
+    } 
 }
