+++ VoiceChat/Assets/VoiceChat/Scripts/Demo/VoiceChatUi.cs
@@ -2,58 +2,61 @@ using UnityEngine;
 using System.Collections;
 using System.Linq;
 
-public class VoiceChatUi : MonoBehaviour
+namespace VoiceChat.Demo.LegacyNetworking
 {
-    void Start()
+    public class VoiceChatUi : MonoBehaviour
     {
-        Application.RequestUserAuthorization(UserAuthorization.Microphone);
-    }
-
-    void OnGUI()
-    {
-        GUILayout.Window(1, new Rect(10, 10, Screen.width - 20, Screen.height - 20), Window, "", GUIStyle.none);
-    }
-
-    void Window(int id)
-    {
-        GUI.Box(new Rect(0, 0, Screen.width - 20, Screen.height - 20), "");
-
-        if (VoiceChatRecorder.Instance.IsRecording)
+        void Start()
         {
-            GUILayout.Label(VoiceChatRecorder.Instance.Device);
-
-            if (GUILayout.Button("Stop Recording"))
-            {
-                VoiceChatRecorder.Instance.StopRecording();
-            }
+            Application.RequestUserAuthorization(UserAuthorization.Microphone);
         }
-        else
-        {
-            GUILayout.Label("Select microphone to start recording");
 
-            foreach (string device in VoiceChatRecorder.Instance.AvailableDevices)
+        void OnGUI()
+        {
+            GUILayout.Window(1, new Rect(10, 10, Screen.width - 20, Screen.height - 20), Window, "", GUIStyle.none);
+        }
+
+        void Window(int id)
+        {
+            GUI.Box(new Rect(0, 0, Screen.width - 20, Screen.height - 20), "");
+
+            if (VoiceChatRecorder.Instance.IsRecording)
             {
-                if (GUILayout.Button(device))
+                GUILayout.Label(VoiceChatRecorder.Instance.Device);
+
+                if (GUILayout.Button("Stop Recording"))
                 {
-                    VoiceChatRecorder.Instance.Device = device;
-                    VoiceChatRecorder.Instance.StartRecording();
+                    VoiceChatRecorder.Instance.StopRecording();
                 }
             }
-        }
-
-        if (VoiceChatRecorder.Instance.Device != null)
-        {
-            GUILayout.Label("Push-to-talk key: " + VoiceChatRecorder.Instance.PushToTalkKey);
-            GUILayout.Label("Toggle-to-talk key: " + VoiceChatRecorder.Instance.ToggleToTalkKey);
-            GUILayout.Label("Auto detect speech: " + (VoiceChatRecorder.Instance.AutoDetectSpeech ? "On" : "Off"));
-
-            if (GUILayout.Button("Toggle Auto Detect"))
+            else
             {
-                VoiceChatRecorder.Instance.AutoDetectSpeech = !VoiceChatRecorder.Instance.AutoDetectSpeech;
+                GUILayout.Label("Select microphone to start recording");
+
+                foreach (string device in VoiceChatRecorder.Instance.AvailableDevices)
+                {
+                    if (GUILayout.Button(device))
+                    {
+                        VoiceChatRecorder.Instance.Device = device;
+                        VoiceChatRecorder.Instance.StartRecording();
+                    }
+                }
             }
 
-            GUI.color = VoiceChatRecorder.Instance.IsTransmitting ? Color.green : Color.red;
-            GUILayout.Label(VoiceChatRecorder.Instance.IsTransmitting ? "Transmitting" : "Silent");
+            if (VoiceChatRecorder.Instance.Device != null)
+            {
+                GUILayout.Label("Push-to-talk key: " + VoiceChatRecorder.Instance.PushToTalkKey);
+                GUILayout.Label("Toggle-to-talk key: " + VoiceChatRecorder.Instance.ToggleToTalkKey);
+                GUILayout.Label("Auto detect speech: " + (VoiceChatRecorder.Instance.AutoDetectSpeech ? "On" : "Off"));
+
+                if (GUILayout.Button("Toggle Auto Detect"))
+                {
+                    VoiceChatRecorder.Instance.AutoDetectSpeech = !VoiceChatRecorder.Instance.AutoDetectSpeech;
+                }
+
+                GUI.color = VoiceChatRecorder.Instance.IsTransmitting ? Color.green : Color.red;
+                GUILayout.Label(VoiceChatRecorder.Instance.IsTransmitting ? "Transmitting" : "Silent");
+            }
         }
-    }
+    } 
 }
