+++ VoiceChat/Assets/VoiceChat/Scripts/Demo/VoiceChatUnityServer.cs
@@ -1,20 +1,23 @@
 using UnityEngine;
 using System.Collections;
 
-public class VoiceChatUnityServer : MonoBehaviour
+namespace VoiceChat.Demo.LegacyNetworking
 {
-    public int Port = 15000;
-    public int MaxConnections = 8;
-
-    void Start()
+    public class VoiceChatUnityServer : MonoBehaviour
     {
-        Network.InitializeServer(MaxConnections, Port, false);
-        MonoBehaviour.Destroy(GetComponent<VoiceChatUnityClient>());
-    }
+        public int Port = 15000;
+        public int MaxConnections = 8;
 
-    void OnPlayerDisconnected(NetworkPlayer player)
-    {
-        Network.RemoveRPCs(player);
-        Network.DestroyPlayerObjects(player);
-    }
+        void Start()
+        {
+            Network.InitializeServer(MaxConnections, Port, false);
+            MonoBehaviour.Destroy(GetComponent<VoiceChatUnityClient>());
+        }
+
+        void OnPlayerDisconnected(NetworkPlayer player)
+        {
+            Network.RemoveRPCs(player);
+            Network.DestroyPlayerObjects(player);
+        }
+    } 
 }
