+++ VoiceChat/Assets/VoiceChat/Scripts/Editor/VoiceChatRecorderEditor.cs
@@ -3,56 +3,59 @@ using System.Linq;
 using UnityEditor;
 using UnityEngine;
 
-[CustomEditor(typeof(VoiceChatRecorder))]
-public class VoiceChatRecorderEditor : Editor
+namespace VoiceChat
 {
-    public override void OnInspectorGUI()
+    [CustomEditor(typeof(VoiceChatRecorder))]
+    public class VoiceChatRecorderEditor : Editor
     {
-        VoiceChatRecorder recorder = target as VoiceChatRecorder;
-
-        if (Application.isPlaying && Network.isClient)
+        public override void OnInspectorGUI()
         {
-            EditorGUILayout.LabelField("Available Devices", EditorStyles.boldLabel);
+            VoiceChatRecorder recorder = target as VoiceChatRecorder;
 
-            foreach (string device in recorder.AvailableDevices)
+            if (Application.isPlaying && Network.isClient)
             {
-                EditorGUILayout.LabelField(device);
-            }
+                EditorGUILayout.LabelField("Available Devices", EditorStyles.boldLabel);
 
-            EditorGUILayout.LabelField("Selected Device", EditorStyles.boldLabel);
-
-            int index = Mathf.Clamp(Array.IndexOf(recorder.AvailableDevices, recorder.Device), 0, Int32.MaxValue) + 1;
-            string[] devices = new string[1] { "Default" }.Concat(recorder.AvailableDevices).ToArray();
-
-            index = EditorGUILayout.Popup(index, devices);
-
-            if (index != 0)
-            {
-                recorder.Device = devices[index];
-            }
-            else
-            {
-                recorder.Device = null;
-            }
-
-            if (recorder.IsRecording)
-            {
-                if (GUILayout.Button("Stop Recording"))
+                foreach (string device in recorder.AvailableDevices)
                 {
-                    recorder.StopRecording();
+                    EditorGUILayout.LabelField(device);
+                }
+
+                EditorGUILayout.LabelField("Selected Device", EditorStyles.boldLabel);
+
+                int index = Mathf.Clamp(Array.IndexOf(recorder.AvailableDevices, recorder.Device), 0, Int32.MaxValue) + 1;
+                string[] devices = new string[1] { "Default" }.Concat(recorder.AvailableDevices).ToArray();
+
+                index = EditorGUILayout.Popup(index, devices);
+
+                if (index != 0)
+                {
+                    recorder.Device = devices[index];
+                }
+                else
+                {
+                    recorder.Device = null;
+                }
+
+                if (recorder.IsRecording)
+                {
+                    if (GUILayout.Button("Stop Recording"))
+                    {
+                        recorder.StopRecording();
+                    }
+                }
+                else
+                {
+                    if (GUILayout.Button("Start Recording"))
+                    {
+                        recorder.StartRecording();
+                    }
                 }
             }
             else
             {
-                if (GUILayout.Button("Start Recording"))
-                {
-                    recorder.StartRecording();
-                }
+                DrawDefaultInspector();
             }
         }
-        else
-        {
-            DrawDefaultInspector();
-        }
-    }
+    } 
 }
