+++ VoiceChat/Assets/VoiceChat/Scripts/Network/VoiceChatNetworkProxy.cs
@@ -1,95 +1,98 @@
 using UnityEngine;
 using System.Collections.Generic;
 
-public class VoiceChatNetworkProxy : MonoBehaviour
+namespace VoiceChat.Networking.Legacy
 {
-    static int networkIdCounter = 0;
-
-    int assignedNetworkId = -1;
-    VoiceChatPlayer player = null;
-    Queue<VoiceChatPacket> packets = new Queue<VoiceChatPacket>(16);
-
-    void Start()
+    public class VoiceChatNetworkProxy : MonoBehaviour
     {
-        if (GetComponent<NetworkView>().isMine)
+        static int networkIdCounter = 0;
+
+        int assignedNetworkId = -1;
+        VoiceChatPlayer player = null;
+        Queue<VoiceChatPacket> packets = new Queue<VoiceChatPacket>(16);
+
+        void Start()
         {
-            VoiceChatRecorder.Instance.NewSample += new System.Action<VoiceChatPacket>(OnNewSample);
-        }
-
-        if (Network.isServer)
-        {
-            assignedNetworkId = ++networkIdCounter;
-            GetComponent<NetworkView>().RPC("SetNetworkId", GetComponent<NetworkView>().owner, assignedNetworkId);
-        }
-
-        if(Network.isClient && (!GetComponent<NetworkView>().isMine || VoiceChatSettings.Instance.LocalDebug))
-        {
-            gameObject.AddComponent<AudioSource>();
-            player = gameObject.AddComponent<VoiceChatPlayer>();
-        }
-    }
-
-    void OnNewSample(VoiceChatPacket packet)
-    {
-        packets.Enqueue(packet);
-    }
-
-    [RPC]
-    void SetNetworkId(int networkId)
-    {
-        VoiceChatRecorder.Instance.NetworkId = networkId;
-    }
-
-    void OnSerializeNetworkView(BitStream stream, NetworkMessageInfo info)
-    {
-        int count = packets.Count;
-
-        if (stream.isWriting)
-        {
-            stream.Serialize(ref count);
-
-            while (packets.Count > 0)
+            if (GetComponent<NetworkView>().isMine)
             {
-                VoiceChatPacket packet = packets.Dequeue();
-                stream.WritePacket(packet);
+                VoiceChatRecorder.Instance.NewSample += OnNewSample;
+            }
 
-                // If this packet is the same size as the sample size, we can return it
-                if (packet.Data.Length == VoiceChatSettings.Instance.SampleSize)
-                {
-                    VoiceChatBytePool.Instance.Return(packet.Data);
-                }
+            if (Network.isServer)
+            {
+                assignedNetworkId = ++networkIdCounter;
+                GetComponent<NetworkView>().RPC("SetNetworkId", GetComponent<NetworkView>().owner, assignedNetworkId);
+            }
+
+            if (Network.isClient && (!GetComponent<NetworkView>().isMine || VoiceChatSettings.Instance.LocalDebug))
+            {
+                gameObject.AddComponent<AudioSource>();
+                player = gameObject.AddComponent<VoiceChatPlayer>();
             }
         }
-        else
+
+        void OnNewSample(VoiceChatPacket packet)
         {
-            if (Network.isServer)
+            packets.Enqueue(packet);
+        }
+
+        [RPC]
+        void SetNetworkId(int networkId)
+        {
+            VoiceChatRecorder.Instance.NetworkId = networkId;
+        }
+
+        void OnSerializeNetworkView(BitStream stream, NetworkMessageInfo info)
+        {
+            int count = packets.Count;
+
+            if (stream.isWriting)
             {
                 stream.Serialize(ref count);
 
-                for (int i = 0; i < count; ++i)
+                while (packets.Count > 0)
                 {
-                    packets.Enqueue(stream.ReadPacket());
+                    VoiceChatPacket packet = packets.Dequeue();
+                    stream.WritePacket(packet);
 
-                    if (Network.connections.Length < 2)
+                    // If this packet is the same size as the sample size, we can return it
+                    if (packet.Data.Length == VoiceChatSettings.Instance.SampleSize)
                     {
-                        packets.Dequeue();
+                        VoiceChatBytePool.Instance.Return(packet.Data);
                     }
                 }
             }
             else
             {
-                stream.Serialize(ref count);
-
-                for (int i = 0; i < count; ++i)
+                if (Network.isServer)
                 {
-                    var packet = stream.ReadPacket();
+                    stream.Serialize(ref count);
 
-                    if (player != null)
+                    for (int i = 0; i < count; ++i)
                     {
-                        player.OnNewSample(packet);
+                        packets.Enqueue(stream.ReadPacket());
+
+                        if (Network.connections.Length < 2)
+                        {
+                            packets.Dequeue();
+                        }
+                    }
+                }
+                else
+                {
+                    stream.Serialize(ref count);
+
+                    for (int i = 0; i < count; ++i)
+                    {
+                        var packet = stream.ReadPacket();
+
+                        if (player != null)
+                        {
+                            player.OnNewSample(packet);
+                        }
                     }
                 }
             }
         }
-    }
+    } 
 }
