+++ VoiceChat/Assets/VoiceChat/Scripts/Network/VoiceChatNetworkUtils.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using UnityEngine;
+
+namespace VoiceChat.Networking.Legacy
+{
+    class VoiceChatNetworkUtils
+    {
+        public static VoiceChatNetworkProxy CreateProxy()
+        {
+            if (!Network.isClient)
+            {
+                Debug.LogError("You're not a client in the unity networking");
+                return null;
+            }
+
+            GameObject prefab = Resources.Load("VoiceChat_NetworkProxy") as GameObject;
+            GameObject instance = Network.Instantiate(prefab, Vector3.zero, Quaternion.identity, 0) as GameObject;
+            return instance.GetComponent<VoiceChatNetworkProxy>();
+        }
+    }
+}
