+++ VoiceChat/Assets/VoiceChat/Scripts/Network/VoiceChatNetworkUtils.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 0c653830dcf591d4d8b3b842a5aadeb5
+timeCreated: 1435253429
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
