+++ VoiceChat/Assets/VoiceChat/Scripts/VoiceChatPlayer.cs
@@ -1,126 +1,129 @@
 using System;
 using UnityEngine;
 
-[RequireComponent(typeof(AudioSource))]
-public class VoiceChatPlayer : MonoBehaviour
+namespace VoiceChat
 {
-    float lastTime = 0;
-    double played = 0;
-    double received = 0;
-    int index = 0;
-    float[] data;
-    float playDelay = 0;
-    bool shouldPlay = false;
-    float lastRecvTime = 0;
-    NSpeex.SpeexDecoder speexDec = new NSpeex.SpeexDecoder(NSpeex.BandMode.Narrow);
-
-    [SerializeField]
-    int playbackDelay = 2;
-
-    public float LastRecvTime
+    [RequireComponent(typeof(AudioSource))]
+    public class VoiceChatPlayer : MonoBehaviour
     {
-        get { return lastRecvTime; }
-    }
+        float lastTime = 0;
+        double played = 0;
+        double received = 0;
+        int index = 0;
+        float[] data;
+        float playDelay = 0;
+        bool shouldPlay = false;
+        float lastRecvTime = 0;
+        NSpeex.SpeexDecoder speexDec = new NSpeex.SpeexDecoder(NSpeex.BandMode.Narrow);
 
-    void Start()
-    {
-        int size = VoiceChatSettings.Instance.Frequency * 10;
+        [SerializeField]
+        int playbackDelay = 2;
 
-        GetComponent<AudioSource>().loop = true;
-        GetComponent<AudioSource>().clip = AudioClip.Create("VoiceChat", size, 1, VoiceChatSettings.Instance.Frequency, false);
-        data = new float[size];
-
-        if (VoiceChatSettings.Instance.LocalDebug)
+        public float LastRecvTime
         {
-            VoiceChatRecorder.Instance.NewSample += OnNewSample;
+            get { return lastRecvTime; }
         }
-    }
 
-    void Update()
-    {
-        if (GetComponent<AudioSource>().isPlaying)
+        void Start()
         {
-            // Wrapped around
-            if (lastTime > GetComponent<AudioSource>().time)
-            {
-                played += GetComponent<AudioSource>().clip.length;
-            }
+            int size = VoiceChatSettings.Instance.Frequency * 10;
 
-            lastTime = GetComponent<AudioSource>().time;
+            GetComponent<AudioSource>().loop = true;
+            GetComponent<AudioSource>().clip = AudioClip.Create("VoiceChat", size, 1, VoiceChatSettings.Instance.Frequency, false);
+            data = new float[size];
 
-            // Check if we've played to far
-            if (played + GetComponent<AudioSource>().time >= received)
+            if (VoiceChatSettings.Instance.LocalDebug)
             {
-                Stop();
-                shouldPlay = false;
+                VoiceChatRecorder.Instance.NewSample += OnNewSample;
             }
         }
-        else
-        {
-            if (shouldPlay)
-            {
-                playDelay -= Time.deltaTime;
 
-                if (playDelay <= 0)
+        void Update()
+        {
+            if (GetComponent<AudioSource>().isPlaying)
+            {
+                // Wrapped around
+                if (lastTime > GetComponent<AudioSource>().time)
                 {
-                    GetComponent<AudioSource>().Play();
+                    played += GetComponent<AudioSource>().clip.length;
+                }
+
+                lastTime = GetComponent<AudioSource>().time;
+
+                // Check if we've played to far
+                if (played + GetComponent<AudioSource>().time >= received)
+                {
+                    Stop();
+                    shouldPlay = false;
+                }
+            }
+            else
+            {
+                if (shouldPlay)
+                {
+                    playDelay -= Time.deltaTime;
+
+                    if (playDelay <= 0)
+                    {
+                        GetComponent<AudioSource>().Play();
+                    }
                 }
             }
         }
-    }
 
-    void Stop()
-    {
-        GetComponent<AudioSource>().Stop();
-        GetComponent<AudioSource>().time = 0;
-        index = 0;
-        played = 0;
-        received = 0;
-        lastTime = 0;
-    }
-
-    public void OnNewSample(VoiceChatPacket packet)
-    {
-        // Store last packet
-
-        // Set last time we got something
-        lastRecvTime = Time.time;
-
-        // Decompress
-        float[] sample = null;
-        int length = VoiceChatUtils.Decompress(speexDec, packet, out sample);
-
-        // Add more time to received
-        received += VoiceChatSettings.Instance.SampleTime;
-
-        // Push data to buffer
-        Array.Copy(sample, 0, data, index, length);
-
-        // Increase index
-        index += length;
-
-        // Handle wrap-around
-        if (index >= GetComponent<AudioSource>().clip.samples)
+        void Stop()
         {
+            GetComponent<AudioSource>().Stop();
+            GetComponent<AudioSource>().time = 0;
             index = 0;
+            played = 0;
+            received = 0;
+            lastTime = 0;
         }
 
-        // Set data
-        GetComponent<AudioSource>().clip.SetData(data, 0);
-
-        // If we're not playing
-        if (!GetComponent<AudioSource>().isPlaying)
+        public void OnNewSample(VoiceChatPacket packet)
         {
-            // Set that we should be playing
-            shouldPlay = true;
+            // Store last packet
 
-            // And if we have no delay set, set it.
-            if (playDelay <= 0)
+            // Set last time we got something
+            lastRecvTime = Time.time;
+
+            // Decompress
+            float[] sample = null;
+            int length = VoiceChatUtils.Decompress(speexDec, packet, out sample);
+
+            // Add more time to received
+            received += VoiceChatSettings.Instance.SampleTime;
+
+            // Push data to buffer
+            Array.Copy(sample, 0, data, index, length);
+
+            // Increase index
+            index += length;
+
+            // Handle wrap-around
+            if (index >= GetComponent<AudioSource>().clip.samples)
             {
-                playDelay = (float)VoiceChatSettings.Instance.SampleTime * playbackDelay;
+                index = 0;
             }
-        }
 
-        VoiceChatFloatPool.Instance.Return(sample);
-    }
+            // Set data
+            GetComponent<AudioSource>().clip.SetData(data, 0);
+
+            // If we're not playing
+            if (!GetComponent<AudioSource>().isPlaying)
+            {
+                // Set that we should be playing
+                shouldPlay = true;
+
+                // And if we have no delay set, set it.
+                if (playDelay <= 0)
+                {
+                    playDelay = (float)VoiceChatSettings.Instance.SampleTime * playbackDelay;
+                }
+            }
+
+            VoiceChatFloatPool.Instance.Return(sample);
+        }
+    } 
 }
